services/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

# DBs / stores
psycopg[binary]==3.1.18
minio==7.2.7
qdrant-client==1.9.1
neo4j==5.22.0

# utils
numpy==1.26.4
pydantic==2.8.2
python-dotenv==1.0.1

PyYAML==6.0.2

geoip2==5.1.0

scikit-learn==1.4.2
joblib==1.4.2

xgboost>=1.7.6